🚚 app.rs から panels.rs（About/支援ページ/自動起動プロンプト）を切り出す
src/app.rs を src/app/ ディレクトリ化し、prompt_login_item_if_needed /
show_about_panel / about_credits / open_support_page とその定数を
panels.rs へ移した。責務が独立していて他のロジックと絡まないため、
extern "C" のシム化はせず丸ごと移動している。
